change structure of data map
